Abstract Piece Changes
Added abstract methods for getting each pieces type as well as the list of valid moves. Constructor now takes a boolean to keep track of color and the initial position of the piece. Default is -1 for x and y to indicate the piece if not on the board since 0,0 is a valid location..